Extract tag display component from tag display page
